requirements.txt
# Essential dependencies for book_creator
python-dotenv>=0.21.0
smolagents>=0.1.0
pillow>=10.0.0
openai>=1.0.0
litellm>=1.0.0

# Optional dependencies
requests>=2.30.0
numpy>=1.26.0 